requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
Faker==25.3.0
fastjsonschema==2.19.1
idna==3.7
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.45
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyproject_hooks==1.1.0
pytest==8.2.1
python-dateutil==2.9.0.post0
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
yarg==0.1.9
